+++ mrb_doc/shaders.rb
@@ -1,7 +1,8 @@
 # Loads a shader from a file
-# @param path [String]
+# @param vertex_shader_path [String]
+# @param fragement_shader_path [String]
 # @return [Shader]
-def load_shader(path)
+def load_shader(vertex_shader_path, fragment_shader_path)
   # mrb_load_shader
   # src/mruby_integration/shaders.cpp
   Shader.new
@@ -14,4 +15,13 @@ def unload_shader(shader)
   # mrb_unload_shader
   # src/mruby_integration/shaders.cpp
   nil
+end
+
+# Checks if the shader is ready
+# @param shader [Shader]
+# @return [Boolean]
+def shader_ready?(shader)
+  # mrb_shader_ready
+  # src/mruby_integration/shaders.cpp
+  True
 end