+++ test/models/shader_test.rb
@@ -15,6 +15,18 @@ class Test
         assert_equal 5, shader.id
 
       end
+
+      def test_shader_ready
+        skip_unless_display_present
+
+        set_window_title(__method__.to_s)
+
+        shader = Shader.load('', "assets/fragment_shader_#{GLSL_VERSION}.fs")
+
+        assert_true shader.ready?
+
+        shader.unload
+      end
     end
   end
 end