+++ test/shader_test.rb
@@ -16,5 +16,17 @@ class Test
 
       unload_shader(shader)
     end
+
+    def test_shader_ready
+      skip_unless_display_present
+
+      set_window_title(__method__.to_s)
+
+      shader = load_shader('', "assets/fragment_shader_#{GLSL_VERSION}.fs")
+
+      assert_true shader_ready?(shader)
+
+      unload_shader(shader)
+    end
   end
 end